refactor: consolidated all github_ workspaces into `github`
