+++ packages/github/src/client/GitHub.mjs
@@ -1,7 +1,11 @@
 import * as core from "@actions/core";
 import { Octokit } from "@octokit/rest";
-import { MAX_ISSUES_LIMIT, MAX_ISSUES_PER_PAGE } from "github_issues/src/constants.mjs";
-import { MAX_PULL_REQUESTS_LIMIT, MAX_PULL_REQUESTS_PER_PAGE } from "github_pull_requests/src/constants.mjs";
+import {
+    MAX_ISSUES_LIMIT,
+    MAX_ISSUES_PER_PAGE,
+    MAX_PULL_REQUESTS_LIMIT,
+    MAX_PULL_REQUESTS_PER_PAGE
+} from "packages/github/src/constants.mjs";
 import { z } from "zod";
 import { issueSchema, pullRequestAsIssueSchema } from "../schemas/issues.mjs";
 import { pullRequestSchema } from "../schemas/pull_requests.js";
