--- packages/github_actions/package.json
@@ -1,15 +1,0 @@
-{
-	"name": "github_actions",
-	"version": "1.0.0",
-	"description": "",
-	"main": "index.js",
-	"directories": {
-		"test": "tests"
-	},
-	"devDependencies": {},
-	"scripts": {
-		"test": "echo \"Error: no test specified\" && exit 1"
-	},
-	"author": "",
-	"license": "ISC"
-}
