--- packages/github_issues/package.json
@@ -1,11 +1,0 @@
-{
-	"name": "github_issues",
-	"version": "1.0.0",
-	"description": "GitHub Issues management",
-	"devDependencies": {},
-	"scripts": {
-		"test": "vitest"
-	},
-	"author": "",
-	"license": "ISC"
-}
