--- packages/github_pull_requests/package.json
@@ -1,14 +1,0 @@
-{
-	"name": "github_pull_requests",
-	"version": "1.0.0",
-	"description": "GitHub Pull Requests management",
-	"directories": {
-		"test": "tests"
-	},
-	"devDependencies": {},
-	"scripts": {
-		"test": "vitest"
-	},
-	"author": "",
-	"license": "ISC"
-}
